chore(core-interactive-icon): rename components for new section in docs
