+++ packages/InteractiveIcon/svgs/CaretDown.jsx
@@ -1,6 +1,6 @@
 import React, { forwardRef } from 'react'
 
-import LimitedInteractiveIcon, { StyledLimitedInteractiveIconSVG } from '../LimitedInteractiveIcon'
+import LimitedInteractiveIcon, { StyledLimitedInteractiveIconSVG } from '../Limited'
 
 const CaretDown = forwardRef((props, ref) => (
   <LimitedInteractiveIcon {...props} ref={ref}>
